Document Lib further and move abc's
